+++ changelog.md
@@ -1,10 +1,10 @@
 # 1.4.0
 - Internal: Reorganize code.
 - UI: Add a "never show again" option in the `lua.dll` popup.
-- UI: Added new select UI for plugins, scripts, and downloading new plugins.
+- UI: Added new select UI for plugins and scripts.
 - UI: Removed plugin/script button in scripts/plugins UI.
 - UI: Unnoticeably faster plugin/script lookup.
-- Plugin Index: Added new plugin index for installing plugins in-game.
+- Plugins: Install plugin via URL.
 - Script termination: Log error properly (how did it take me this long to find this)
 
 # 1.3.0